restructure files according to new standard
