+++ src/signals/incident-management/containers/ReporterContainer/components/ContactHistory.test.tsx
@@ -2,16 +2,16 @@
 // Copyright (C) 2021 Gemeente Amsterdam
 import { render, screen, waitFor } from '@testing-library/react'
 import * as reactRedux from 'react-redux'
+
 import { showGlobalNotification } from 'containers/App/actions'
 import { withAppContext } from 'test/utils'
 
+import * as API from '../../../../../../internals/testing/api'
 import {
   fetchMock,
   mockRequestHandler,
-} from '../../../../../../../internals/testing/msw-server'
-import * as API from '../../../../../../../internals/testing/api'
-
-import ContactHistory from '../ContactHistory'
+} from '../../../../../../internals/testing/msw-server'
+import ContactHistory from './ContactHistory'
 
 const dispatch = jest.fn()
 jest.spyOn(reactRedux, 'useDispatch').mockImplementation(() => dispatch)
