+++ src/signals/incident-management/containers/ReporterContainer/components/Header.test.tsx
@@ -1,9 +1,10 @@
 // SPDX-License-Identifier: MPL-2.0
 // Copyright (C) 2021 Gemeente Amsterdam
 import { render, screen } from '@testing-library/react'
+
 import { withAppContext } from 'test/utils'
 
-import Header from '../Header'
+import Header from './Header'
 
 describe('Header', () => {
   it('renders correctly', () => {
