+++ src/signals/incident-management/containers/ReporterContainer/components/Header.tsx
@@ -1,10 +1,12 @@
 // SPDX-License-Identifier: MPL-2.0
 // Copyright (C) 2021 Gemeente Amsterdam
-import { Heading, themeSpacing } from '@amsterdam/asc-ui'
-import { INCIDENT_URL } from 'signals/incident-management/routes'
-import BackLink from 'components/BackLink'
-import styled from 'styled-components'
 import type { FunctionComponent } from 'react'
+
+import { Heading, themeSpacing } from '@amsterdam/asc-ui'
+import styled from 'styled-components'
+
+import BackLink from 'components/BackLink'
+import { INCIDENT_URL } from 'signals/incident-management/routes'
 
 const StyledHeading = styled(Heading)`
   padding-top: ${themeSpacing(6)};
