+++ src/signals/incident-management/containers/ReporterContainer/components/IncidentList.test.tsx
@@ -2,9 +2,9 @@
 // Copyright (C) 2021 Gemeente Amsterdam
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import type { ReporterIncident } from '../../types'
 
-import IncidentList from '../IncidentList'
+import type { ReporterIncident } from '../types'
+import IncidentList from './IncidentList'
 
 describe('IncidentList', () => {
   const list: ReporterIncident[] = [
